requirements.txt
Pillow>=10.0.0
jpholiday>=1.0.0